rename app and remove celery from code
