--- src/import_export_celery/locale/eu/LC_MESSAGES/django.po
@@ -1,149 +1,0 @@
-# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER
-# This file is distributed under the same license as the PACKAGE package.
-#
-# Urtzi Odriozola <uodriozola@codesyntax.com>, 2023.
-msgid ""
-msgstr ""
-"Project-Id-Version: \n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-05-17 20:55+0000\n"
-"PO-Revision-Date: 2023-06-07 21:56+0100\n"
-"Last-Translator: Urtzi Odriozola <uodriozola@codesyntax.com>\n"
-"Language-Team: \n"
-"Language: eu\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=2; plural=(n > 1);\n"
-
-#: import_export_celery/admin.py:12
-msgid "Job status info"
-msgstr "Lanaren egoera"
-
-#: import_export_celery/admin.py:22 import_export_celery/models/importjob.py:62
-msgid "Name of model to import to"
-msgstr "Inportatu beharreko modeloaren izena"
-
-#: import_export_celery/admin_actions.py:20
-msgid "Perform import"
-msgstr "Hasi inportazioa"
-
-#: import_export_celery/admin_actions.py:29
-msgid "Perform dry import"
-msgstr "Hasi probako inportazioa"
-
-#: import_export_celery/admin_actions.py:39
-msgid "Run export job"
-msgstr "Hasi esportazio lana"
-
-#: import_export_celery/admin_actions.py:67
-msgid "Export with celery"
-msgstr "Exportatu celery bidez"
-
-#: import_export_celery/apps.py:7
-msgid "Import Export Celery"
-msgstr "Inportatu - Esportatu Celery bidez"
-
-#: import_export_celery/models/exportjob.py:27
-msgid "exported file"
-msgstr "esportatutako fitxategia"
-
-#: import_export_celery/models/exportjob.py:35
-#: import_export_celery/models/importjob.py:30
-msgid "Have we started processing the file? If so when?"
-msgstr "Fitxategia prozesatzen hasi gara? Hala bada, noiz?"
-
-#: import_export_celery/models/exportjob.py:42
-#: import_export_celery/models/importjob.py:67
-msgid "Status of the job"
-msgstr "Lanaren egoera"
-
-#: import_export_celery/models/exportjob.py:48
-msgid "Format of file to be exported"
-msgstr "Esportatu beharreko fitxategiaren formatua"
-
-#: import_export_celery/models/exportjob.py:55
-msgid "App label of model to export from"
-msgstr "Esportatu beharreko modeloaren aplikazio izena"
-
-#: import_export_celery/models/exportjob.py:60
-msgid "Name of model to export from"
-msgstr "Esportatu beharreko modeloaren izena"
-
-#: import_export_celery/models/exportjob.py:65
-msgid "Resource to use when exporting"
-msgstr "Esportatzeko erabili beharreko baliabidea"
-
-#: import_export_celery/models/exportjob.py:71
-msgid "JSON list of pks to export"
-msgstr "Esportatzeko ID-en JSON zerrenda"
-
-#: import_export_celery/models/exportjob.py:76
-msgid "Send me an email when this export job is complete"
-msgstr "Bidali eposta bat esportazio lan hau amaitzen denean"
-
-#: import_export_celery/models/exportjob.py:81
-msgid "Site of origin"
-msgstr "Jatorrizko gunea"
-
-#: import_export_celery/models/exportjob.py:87
-msgid "Export job"
-msgstr "Esportazio lana"
-
-#: import_export_celery/models/exportjob.py:88
-msgid "Export jobs"
-msgstr "Esportazio lanak"
-
-#: import_export_celery/models/importjob.py:22
-msgid "File to be imported"
-msgstr "Inportatu beharreko fitxategia"
-
-#: import_export_celery/models/importjob.py:37
-msgid "Has the import been completed? If so when?"
-msgstr "Inportazioa amaitu da? Hala bada, noiz?"
-
-#: import_export_celery/models/importjob.py:44
-msgid "Format of file to be imported"
-msgstr "Inportatu beharreko fitxategiaren formatua"
-
-#: import_export_celery/models/importjob.py:49
-msgid "Summary of changes made by this import"
-msgstr "Inportazio honek eginiko aldaketen laburpena"
-
-#: import_export_celery/models/importjob.py:56
-msgid "Errors"
-msgstr "Erroreak"
-
-#: import_export_celery/models/importjob.py:73
-msgid "Import job"
-msgstr "Inportazio lana"
-
-#: import_export_celery/models/importjob.py:74
-msgid "Import jobs"
-msgstr "Inportazio lanak"
-
-#: import_export_celery/tasks.py:61
-#, python-format
-msgid "Imported file has a wrong encoding: %s"
-msgstr "Inportatutako fitxategiak kodetze okerra du: %s"
-
-#: import_export_celery/tasks.py:68
-#, python-format
-msgid "Error reading file: %s"
-msgstr "Errorea fitxategia irakurtzean: %s"
-
-#: import_export_celery/tasks.py:101
-#, python-format
-msgid ""
-"Line: %s - %s\n"
-"\t%s\n"
-"%s"
-msgstr ""
-"Lerroa: %s - %s\n"
-"\t%s\n"
-"%s"
-
-#: import_export_celery/tasks.py:190
-#, python-format
-msgid "Import error %s"
-msgstr "Inportazio errorea %s"
