+++ src/import_export_job/apps.py
@@ -3,5 +3,5 @@ from django.utils.translation import gettext_lazy as _
 
 
 class ImportExportCeleryConfig(AppConfig):
-    name = "import_export_celery"
-    verbose_name = _("Import Export Celery")
+    name = "import_export_job"
+    verbose_name = _("Import Export Job")
