+++ src/import_export_job/model_config.py
@@ -1,14 +1,14 @@
-from celery.utils.log import get_task_logger
+#from celery.utils.log import get_task_logger
 from django.apps import apps
 from import_export.resources import modelresource_factory
 
-log = get_task_logger(__name__)
+#log = get_task_logger(__name__)
 
 
 class ModelConfig:
     def __init__(self, app_label=None, model_name=None, resource=None):
         self.model = apps.get_model(app_label=app_label, model_name=model_name)
-        log.debug(resource)
+        #log.debug(resource)
         if resource:
             self.resource = resource()
         else:
