+++ src/import_export_job/models/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (C) 2019 o.s. Auto*Mat
 
 """Import all models."""
-from import_export_celery.models.importjob import ImportJob
+from import_export_job.models.importjob import ImportJob
 
 __all__ = (ImportJob,)
