+++ src/import_export_job/models/importjob.py
@@ -2,26 +2,29 @@
 
 import logging
 
-from author.decorators import with_author
-from django.conf import settings
 from django.db import models, transaction
+from django.contrib.auth.models import User
+
+from django.conf import settings
 from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
-from import_export_celery.fields import ImportExportFileField
-from import_export_celery.tasks import run_import_job
-from import_export_celery.utils import DEFAULT_FORMATS
+
+from import_export_job.tasks import run_import_job
+from import_export_job.utils import DEFAULT_FORMATS
 
 logger = logging.getLogger(__name__)
 
 
-@with_author
 class ImportJob(models.Model):
-    file = ImportExportFileField(
+    updated_at = models.DateTimeField(auto_now=True)
+    owner = models.ForeignKey(User, null=True, blank=True, on_delete=models.SET_NULL)
+    
+    file = models.FileField(
         verbose_name=_("File to be imported"),
-        upload_to="django-import-export-celery-import-jobs",
+        upload_to="django-import-jobs/",
         blank=False,
         null=False,
         max_length=255,
@@ -46,9 +49,9 @@ class ImportJob(models.Model):
         max_length=255,
     )
 
-    change_summary = ImportExportFileField(
+    change_summary = models.FileField(
         verbose_name=_("Summary of changes made by this import"),
-        upload_to="django-import-export-celery-import-change-summaries",
+        upload_to="django-import-job-change-summaries/",
         blank=True,
         null=True,
     )
